Move all commands to main project and delete `ScorePump`
