--- osu.Server.Queues.ScorePump/Dockerfile
@@ -1,14 +1,0 @@
-FROM mcr.microsoft.com/dotnet/sdk:6.0 AS build-env
-WORKDIR /app
-
-# Copy everything else and build
-COPY . ./
-RUN dotnet publish osu.Server.Queues.ScorePump -c Release -o out
-# get rid of bloat
-RUN rm -rf ./out/runtimes ./out/osu.Game.Resources.dll
-
-# Build runtime image
-FROM mcr.microsoft.com/dotnet/runtime:6.0
-WORKDIR /app
-COPY --from=build-env /app/out .
-ENTRYPOINT ["dotnet", "osu.Server.Queues.ScorePump.dll"]
