--- osu.Server.Queues.ScorePump/osu.Server.Queues.ScorePump.csproj
@@ -1,21 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>net6.0</TargetFramework>
-        <Nullable>enable</Nullable>
-        <RootNamespace>osu.Server.Queues.ScorePump</RootNamespace>
-        <AssemblyName>osu.Server.Queues.ScorePump</AssemblyName>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Dapper" Version="2.0.123" />
-        <PackageReference Include="McMaster.Extensions.CommandLineUtils" Version="4.0.2" />
-        <PackageReference Include="ppy.osu.Server.OsuQueueProcessor" Version="2022.1220.0" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\osu.Server.Queues.ScoreStatisticsProcessor\osu.Server.Queues.ScoreStatisticsProcessor.csproj" />
-    </ItemGroup>
-
-</Project>
