+++ osu.Server.Queues.ScoreStatisticsProcessor/Commands/BaseCommand.cs
@@ -1,10 +1,10 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace osu.Server.Queues.ScorePump
+namespace osu.Server.Queues.ScoreStatisticsProcessor.Commands
 {
     public abstract class BaseCommand
     {
-        protected readonly ScoreStatisticsProcessor.ScoreStatisticsProcessor Queue = new ScoreStatisticsProcessor.ScoreStatisticsProcessor();
+        protected readonly ScoreStatisticsProcessor Queue = new ScoreStatisticsProcessor();
     }
 }
