+++ osu.Server.Queues.ScoreStatisticsProcessor/Commands/Performance/PerformanceCommand.cs
@@ -8,10 +8,9 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using McMaster.Extensions.CommandLineUtils;
-using osu.Server.Queues.ScoreStatisticsProcessor;
 using osu.Server.Queues.ScoreStatisticsProcessor.Processors;
 
-namespace osu.Server.Queues.ScorePump.Performance
+namespace osu.Server.Queues.ScoreStatisticsProcessor.Commands.Performance
 {
     public abstract class PerformanceCommand : BaseCommand
     {
