+++ osu.Server.Queues.ScoreStatisticsProcessor/Commands/Performance/UserTotals/UpdateAllUserTotalsCommand.cs
@@ -7,9 +7,8 @@ using System.Threading;
 using System.Threading.Tasks;
 using Dapper;
 using McMaster.Extensions.CommandLineUtils;
-using osu.Server.Queues.ScoreStatisticsProcessor;
 
-namespace osu.Server.Queues.ScorePump.Performance.UserTotals
+namespace osu.Server.Queues.ScoreStatisticsProcessor.Commands.Performance.UserTotals
 {
     [Command("all", Description = "Updates the total PP of all users.")]
     public class UpdateAllUserTotalsCommand : PerformanceCommand
