+++ osu.Server.Queues.ScoreStatisticsProcessor/Commands/QueueCommands.cs
@@ -2,9 +2,9 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using McMaster.Extensions.CommandLineUtils;
-using osu.Server.Queues.ScorePump.Queue;
+using osu.Server.Queues.ScoreStatisticsProcessor.Commands.Queue;
 
-namespace osu.Server.Queues.ScorePump
+namespace osu.Server.Queues.ScoreStatisticsProcessor.Commands
 {
     [Command(Name = "queue", Description = "Perform various operations on the processing queue.")]
     [Subcommand(typeof(PumpTestDataCommand))]
