+++ osu.Server.Queues.ScoreStatisticsProcessor/Program.cs
@@ -1,10 +1,36 @@
-﻿namespace osu.Server.Queues.ScoreStatisticsProcessor
+﻿using System;
+using System.Threading;
+using System.Threading.Tasks;
+using McMaster.Extensions.CommandLineUtils;
+using osu.Server.Queues.ScoreStatisticsProcessor.Commands;
+
+namespace osu.Server.Queues.ScoreStatisticsProcessor
 {
+    [Command]
+    [Subcommand(typeof(QueueCommands))]
+    [Subcommand(typeof(PerformanceCommands))]
+    [Subcommand(typeof(RecalculateScoresCommand))]
     public class Program
     {
-        public static void Main(string[] args)
+        private static readonly CancellationTokenSource cts = new CancellationTokenSource();
+
+        public static async Task Main(string[] args)
         {
-            new ScoreStatisticsProcessor().Run();
+            Console.CancelKeyPress += (_, e) =>
+            {
+                Console.WriteLine("Cancellation requested!");
+                cts.Cancel();
+
+                e.Cancel = true;
+            };
+
+            await CommandLineApplication.ExecuteAsync<Program>(args, cts.Token);
+        }
+
+        public int OnExecute(CommandLineApplication app)
+        {
+            app.ShowHelp();
+            return 1;
         }
     }
 }
