+++ osu.Server.Queues.ScoreStatisticsProcessor/osu.Server.Queues.ScoreStatisticsProcessor.csproj
@@ -11,6 +11,7 @@
     <ItemGroup>
         <PackageReference Include="Dapper" Version="2.0.123" />
         <PackageReference Include="Dapper.Contrib" Version="2.0.78" />
+        <PackageReference Include="McMaster.Extensions.CommandLineUtils" Version="4.0.2" />
         <PackageReference Include="ppy.osu.Game" Version="2023.814.0" />
         <PackageReference Include="ppy.osu.Game.Rulesets.Catch" Version="2023.814.0" />
         <PackageReference Include="ppy.osu.Game.Rulesets.Mania" Version="2023.814.0" />
